pull NABERS offices inputs and calcs back into one file, add ESC calculation parameters and tests, add test of privacy files
