+++ openfisca_nsw_ess_nabers/parameters/energy_saving_scheme/electricity_certificate_conversion_factor.yaml
@@ -5,4 +5,4 @@ reference: Equation 1 of the ESS Rule 2020, located in Clause 6.5 of the Energy
   Supply Act 1995.
 values:
   2020-01-01:
-    value: 1.06
+    value: 1.060
