+++ openfisca_nsw_ess_nabers/tests/nabers/apartments_rev_calculator/NABERS_ratings_tests/AB1173.yaml
@@ -6,7 +6,7 @@
     number_of_apartments: 109
     number_of_central_ac_apartments: 0
     number_of_condenser_water_serviced_apartments: 109
-    pool_input: no_pool
+    pool_status: no_pool
     apartment_has_gym: has_gym
     number_of_naturally_ventilated_parking_spaces: 30
     number_of_mechanically_ventilated_parking_spaces: 0
@@ -23,7 +23,7 @@
     number_of_apartments: 109
     number_of_central_ac_apartments: 0
     number_of_condenser_water_serviced_apartments: 109
-    pool_input: no_pool
+    pool_status: no_pool
     apartment_has_gym: has_gym
     number_of_naturally_ventilated_parking_spaces: 30
     number_of_mechanically_ventilated_parking_spaces: 0
@@ -40,7 +40,7 @@
     number_of_apartments: 109
     number_of_central_ac_apartments: 0
     number_of_condenser_water_serviced_apartments: 109
-    pool_input: no_pool
+    pool_status: no_pool
     apartment_has_gym: has_gym
     number_of_naturally_ventilated_parking_spaces: 30
     number_of_mechanically_ventilated_parking_spaces: 0
@@ -57,7 +57,7 @@
     number_of_apartments: 109
     number_of_central_ac_apartments: 0
     number_of_condenser_water_serviced_apartments: 109
-    pool_input: no_pool
+    pool_status: no_pool
     apartment_has_gym: has_gym
     number_of_naturally_ventilated_parking_spaces: 30
     number_of_mechanically_ventilated_parking_spaces: 0
@@ -74,7 +74,7 @@
     number_of_apartments: 109
     number_of_central_ac_apartments: 0
     number_of_condenser_water_serviced_apartments: 109
-    pool_input: no_pool
+    pool_status: no_pool
     apartment_has_gym: has_gym
     number_of_naturally_ventilated_parking_spaces: 30
     number_of_mechanically_ventilated_parking_spaces: 0
@@ -91,7 +91,7 @@
     number_of_apartments: 109
     number_of_central_ac_apartments: 0
     number_of_condenser_water_serviced_apartments: 109
-    pool_input: no_pool
+    pool_status: no_pool
     apartment_has_gym: has_gym
     number_of_naturally_ventilated_parking_spaces: 30
     number_of_mechanically_ventilated_parking_spaces: 0
@@ -108,7 +108,7 @@
     number_of_apartments: 109
     number_of_central_ac_apartments: 0
     number_of_condenser_water_serviced_apartments: 109
-    pool_input: no_pool
+    pool_status: no_pool
     apartment_has_gym: has_gym
     number_of_naturally_ventilated_parking_spaces: 30
     number_of_mechanically_ventilated_parking_spaces: 0
@@ -126,7 +126,7 @@
     number_of_apartments: 109
     number_of_central_ac_apartments: 0
     number_of_condenser_water_serviced_apartments: 109
-    pool_input: no_pool
+    pool_status: no_pool
     apartment_has_gym: has_gym
     number_of_naturally_ventilated_parking_spaces: 30
     number_of_mechanically_ventilated_parking_spaces: 0
@@ -145,7 +145,7 @@
     number_of_apartments: 109
     number_of_central_ac_apartments: 0
     number_of_condenser_water_serviced_apartments: 109
-    pool_input: no_pool
+    pool_status: no_pool
     apartment_has_gym: has_gym
     number_of_naturally_ventilated_parking_spaces: 30
     number_of_mechanically_ventilated_parking_spaces: 0
@@ -167,7 +167,7 @@
     number_of_apartments: 109
     number_of_central_ac_apartments: 0
     number_of_condenser_water_serviced_apartments: 109
-    pool_input: no_pool
+    pool_status: no_pool
     apartment_has_gym: has_gym
     number_of_naturally_ventilated_parking_spaces: 30
     number_of_mechanically_ventilated_parking_spaces: 0
@@ -189,7 +189,7 @@
     number_of_apartments: 109
     number_of_central_ac_apartments: 0
     number_of_condenser_water_serviced_apartments: 109
-    pool_input: no_pool
+    pool_status: no_pool
     apartment_has_gym: has_gym
     number_of_naturally_ventilated_parking_spaces: 30
     number_of_mechanically_ventilated_parking_spaces: 0
@@ -211,7 +211,7 @@
     number_of_apartments: 109
     number_of_central_ac_apartments: 0
     number_of_condenser_water_serviced_apartments: 109
-    pool_input: no_pool
+    pool_status: no_pool
     apartment_has_gym: has_gym
     number_of_naturally_ventilated_parking_spaces: 30
     number_of_mechanically_ventilated_parking_spaces: 0
@@ -233,7 +233,7 @@
     number_of_apartments: 109
     number_of_central_ac_apartments: 0
     number_of_condenser_water_serviced_apartments: 109
-    pool_input: no_pool
+    pool_status: no_pool
     apartment_has_gym: has_gym
     number_of_naturally_ventilated_parking_spaces: 30
     number_of_mechanically_ventilated_parking_spaces: 0
@@ -255,7 +255,7 @@
     number_of_apartments: 109
     number_of_central_ac_apartments: 0
     number_of_condenser_water_serviced_apartments: 109
-    pool_input: no_pool
+    pool_status: no_pool
     apartment_has_gym: has_gym
     number_of_naturally_ventilated_parking_spaces: 30
     number_of_mechanically_ventilated_parking_spaces: 0
@@ -277,7 +277,7 @@
     number_of_apartments: 109
     number_of_central_ac_apartments: 0
     number_of_condenser_water_serviced_apartments: 109
-    pool_input: no_pool
+    pool_status: no_pool
     apartment_has_gym: has_gym
     number_of_naturally_ventilated_parking_spaces: 30
     number_of_mechanically_ventilated_parking_spaces: 0
@@ -299,7 +299,7 @@
     number_of_apartments: 109
     number_of_central_ac_apartments: 0
     number_of_condenser_water_serviced_apartments: 109
-    pool_input: no_pool
+    pool_status: no_pool
     apartment_has_gym: has_gym
     number_of_naturally_ventilated_parking_spaces: 30
     number_of_mechanically_ventilated_parking_spaces: 0
