+++ openfisca_nsw_ess_nabers/tests/nabers/apartments_rev_calculator/NABERS_ratings_tests/AB1174.yaml
@@ -6,7 +6,7 @@
     number_of_apartments: 164
     number_of_central_ac_apartments: 0
     number_of_condenser_water_serviced_apartments: 164
-    pool_input: heated_pool
+    pool_status: heated_pool
     apartment_has_gym: has_gym
     number_of_naturally_ventilated_parking_spaces: 0
     number_of_mechanically_ventilated_parking_spaces: 84
@@ -23,7 +23,7 @@
     number_of_apartments: 164
     number_of_central_ac_apartments: 0
     number_of_condenser_water_serviced_apartments: 164
-    pool_input: heated_pool
+    pool_status: heated_pool
     apartment_has_gym: has_gym
     number_of_naturally_ventilated_parking_spaces: 0
     number_of_mechanically_ventilated_parking_spaces: 84
@@ -40,7 +40,7 @@
     number_of_apartments: 164
     number_of_central_ac_apartments: 0
     number_of_condenser_water_serviced_apartments: 164
-    pool_input: heated_pool
+    pool_status: heated_pool
     apartment_has_gym: has_gym
     number_of_naturally_ventilated_parking_spaces: 0
     number_of_mechanically_ventilated_parking_spaces: 84
@@ -57,7 +57,7 @@
     number_of_apartments: 164
     number_of_central_ac_apartments: 0
     number_of_condenser_water_serviced_apartments: 164
-    pool_input: heated_pool
+    pool_status: heated_pool
     apartment_has_gym: has_gym
     number_of_naturally_ventilated_parking_spaces: 0
     number_of_mechanically_ventilated_parking_spaces: 84
@@ -74,7 +74,7 @@
     number_of_apartments: 164
     number_of_central_ac_apartments: 0
     number_of_condenser_water_serviced_apartments: 164
-    pool_input: heated_pool
+    pool_status: heated_pool
     apartment_has_gym: has_gym
     number_of_naturally_ventilated_parking_spaces: 0
     number_of_mechanically_ventilated_parking_spaces: 84
@@ -91,7 +91,7 @@
     number_of_apartments: 164
     number_of_central_ac_apartments: 0
     number_of_condenser_water_serviced_apartments: 164
-    pool_input: heated_pool
+    pool_status: heated_pool
     apartment_has_gym: has_gym
     number_of_naturally_ventilated_parking_spaces: 0
     number_of_mechanically_ventilated_parking_spaces: 84
@@ -108,7 +108,7 @@
     number_of_apartments: 164
     number_of_central_ac_apartments: 0
     number_of_condenser_water_serviced_apartments: 164
-    pool_input: heated_pool
+    pool_status: heated_pool
     apartment_has_gym: has_gym
     number_of_naturally_ventilated_parking_spaces: 0
     number_of_mechanically_ventilated_parking_spaces: 84
@@ -125,7 +125,7 @@
     number_of_apartments: 164
     number_of_central_ac_apartments: 0
     number_of_condenser_water_serviced_apartments: 164
-    pool_input: heated_pool
+    pool_status: heated_pool
     apartment_has_gym: has_gym
     number_of_naturally_ventilated_parking_spaces: 0
     number_of_mechanically_ventilated_parking_spaces: 84
@@ -143,7 +143,7 @@
     number_of_apartments: 164
     number_of_central_ac_apartments: 0
     number_of_condenser_water_serviced_apartments: 164
-    pool_input: heated_pool
+    pool_status: heated_pool
     apartment_has_gym: has_gym
     number_of_naturally_ventilated_parking_spaces: 0
     number_of_mechanically_ventilated_parking_spaces: 84
@@ -164,7 +164,7 @@
     number_of_apartments: 164
     number_of_central_ac_apartments: 0
     number_of_condenser_water_serviced_apartments: 164
-    pool_input: heated_pool
+    pool_status: heated_pool
     apartment_has_gym: has_gym
     number_of_naturally_ventilated_parking_spaces: 0
     number_of_mechanically_ventilated_parking_spaces: 84
@@ -185,7 +185,7 @@
     number_of_apartments: 164
     number_of_central_ac_apartments: 0
     number_of_condenser_water_serviced_apartments: 164
-    pool_input: heated_pool
+    pool_status: heated_pool
     apartment_has_gym: has_gym
     number_of_naturally_ventilated_parking_spaces: 0
     number_of_mechanically_ventilated_parking_spaces: 84
@@ -206,7 +206,7 @@
     number_of_apartments: 164
     number_of_central_ac_apartments: 0
     number_of_condenser_water_serviced_apartments: 164
-    pool_input: heated_pool
+    pool_status: heated_pool
     apartment_has_gym: has_gym
     number_of_naturally_ventilated_parking_spaces: 0
     number_of_mechanically_ventilated_parking_spaces: 84
@@ -228,7 +228,7 @@
     number_of_apartments: 164
     number_of_central_ac_apartments: 0
     number_of_condenser_water_serviced_apartments: 164
-    pool_input: heated_pool
+    pool_status: heated_pool
     apartment_has_gym: has_gym
     number_of_naturally_ventilated_parking_spaces: 0
     number_of_mechanically_ventilated_parking_spaces: 84
@@ -251,7 +251,7 @@
     number_of_apartments: 164
     number_of_central_ac_apartments: 0
     number_of_condenser_water_serviced_apartments: 164
-    pool_input: heated_pool
+    pool_status: heated_pool
     apartment_has_gym: has_gym
     number_of_naturally_ventilated_parking_spaces: 0
     number_of_mechanically_ventilated_parking_spaces: 84
@@ -274,7 +274,7 @@
     number_of_apartments: 164
     number_of_central_ac_apartments: 0
     number_of_condenser_water_serviced_apartments: 164
-    pool_input: heated_pool
+    pool_status: heated_pool
     apartment_has_gym: has_gym
     number_of_naturally_ventilated_parking_spaces: 0
     number_of_mechanically_ventilated_parking_spaces: 84
@@ -296,7 +296,7 @@
     number_of_apartments: 164
     number_of_central_ac_apartments: 0
     number_of_condenser_water_serviced_apartments: 164
-    pool_input: heated_pool
+    pool_status: heated_pool
     apartment_has_gym: has_gym
     number_of_naturally_ventilated_parking_spaces: 0
     number_of_mechanically_ventilated_parking_spaces: 84
@@ -318,7 +318,7 @@
     number_of_apartments: 164
     number_of_central_ac_apartments: 0
     number_of_condenser_water_serviced_apartments: 164
-    pool_input: heated_pool
+    pool_status: heated_pool
     apartment_has_gym: has_gym
     number_of_naturally_ventilated_parking_spaces: 0
     number_of_mechanically_ventilated_parking_spaces: 84
@@ -341,7 +341,7 @@
     number_of_apartments: 164
     number_of_central_ac_apartments: 0
     number_of_condenser_water_serviced_apartments: 164
-    pool_input: heated_pool
+    pool_status: heated_pool
     apartment_has_gym: has_gym
     number_of_naturally_ventilated_parking_spaces: 0
     number_of_mechanically_ventilated_parking_spaces: 84
@@ -364,7 +364,7 @@
     number_of_apartments: 164
     number_of_central_ac_apartments: 0
     number_of_condenser_water_serviced_apartments: 164
-    pool_input: heated_pool
+    pool_status: heated_pool
     apartment_has_gym: has_gym
     number_of_naturally_ventilated_parking_spaces: 0
     number_of_mechanically_ventilated_parking_spaces: 84
