+++ openfisca_nsw_ess_nabers/variables/energy_saving_scheme/NABERS_offices/nabers_office_shared_library_of_terms.py
@@ -3,18 +3,18 @@ from openfisca_core.model_api import *
 # Import the Entities specifically defined for this tax and benefit system
 from openfisca_nsw_base.entities import *
 import numpy as np
-
-private = "True"
-if private == "True":
+import math
+private = True
+if private:
     from openfisca_nsw_ess_nabers.variables.energy_saving_scheme.NABERS_offices import coefficient_values as c
 else:
     from openfisca_nsw_ess_nabers.variables.energy_saving_scheme.NABERS_offices import public_coefficient_values as c
-float_formatter = "{:.9f}".format
-np.set_printoptions(formatter={'float_kind':float_formatter})
 
 if not(c.has_real_values):
-    error =  ("real coefficients aren't used, and return values will be incorrect")
-    print (error)
+    error = ("real coefficients aren't used, and return values will be incorrect")
+
+float_formatter = "{:.9f}".format
+np.set_printoptions(formatter={'float_kind':float_formatter})
 
 # measured_electricity_consumption input at Step 1
 # measured_gas_consumption input at Step 1
@@ -35,6 +35,160 @@ if not(c.has_real_values):
 
 # note that this file is used to store variables which are IDENTICAL across
 # NABERS office reverse calculators. This file is to be made PRIVATE.
+
+class offices_postcode:
+    value_type = float
+    entity = Building
+    definition_period = ETERNITY
+    label = 'The postcode for the relevant NABERS building.'
+
+    def formula(buildings, period, parameters):
+        return (buildings('postcode', period))
+
+
+class number_of_computers(Variable):
+    value_type = float
+    entity = Building
+    definition_period = ETERNITY
+    label = "The number of computers registered as used within the NABERS rating"
+
+
+class rating_type(Variable):
+    value_type = str
+    entity = Building
+    definition_period = ETERNITY
+    label = 'The type of building rated within the NABERS Office suite. Also' \
+            ' used to determine which reverse calculator is used.'
+
+
+class benchmark_star_rating(Variable):
+    value_type = float
+    entity = Building
+    definition_period = ETERNITY  # need to check whether these inputs, on the NABERS reports, should all be year
+    label = 'The star rating for which the benchmark electricity and gas' \
+            'consumption is calculated against - what NABERS rating the' \
+            ' building aims to achieve.'
+
+    def formula(buildings, period, parameters):
+        method_one = buildings('method_one', period)
+        method_two = buildings('method_two', period)
+        condition_method_one = buildings('method_one_can_be_used', period) == True
+        return where (condition_method_one, method_one, method_two)
+
+
+class building_state_location(Variable):
+    value_type = str
+    entity = Building
+    definition_period = ETERNITY
+    label = "State within which the relevant NABERS rated building is located."
+
+
+class elec_kWh(Variable):
+    value_type = float
+    entity = Building
+    definition_period = ETERNITY
+    label = "The electricity consumption of the building, in kWh, as detailed on the NABERS report"
+
+
+class gas_kWh(Variable):
+    value_type = float
+    entity = Building
+    definition_period = ETERNITY
+    label = "The gas consumption of the building, in kWh, as detailed on the NABERS report"
+
+    def formula(buildings, period, parameters):
+        return buildings('gas_MJ_to_KWh', period)
+
+
+class diesel_kWh(Variable):
+    value_type = float
+    entity = Building
+    definition_period = ETERNITY
+    label = "The oil consumption of the building, in kWh, as detailed on the NABERS report"
+
+    def formula(buildings, period, parameters):
+        return buildings('diesel_litres_to_KWh', period)
+
+
+class coal_kWh (Variable):
+    value_type = float
+    entity = Building
+    definition_period = ETERNITY
+    label = "The coal consumption of the building, in kWh, as detailed on the NABERS report"
+
+    def formula(buildings, period, parameters):
+        return buildings('coal_KG_to_KWh', period)
+
+
+class total_energy_kwh(Variable):
+    value_type = float
+    entity = Building
+    definition_period = ETERNITY
+    label = "The total kWh consumption of the building, summing electricity, gas, oil and diesel, expressed in kWh"
+
+    def formula(buildings, period, parameters):
+        return buildings('elec_kWh', period) + buildings('gas_kWh', period) + buildings('coal_kWh', period) + buildings('diesel_kWh', period)
+
+
+class perc_elec_kwh(Variable):
+    value_type = float
+    entity = Building
+    definition_period = ETERNITY
+    label = "The percentage of a building's energy consumption that is derived from its electricity consumption, expressed as a percentage of total energy use to 2 decimal places"
+
+    def formula(buildings, period, parameters):
+        elec_percent = buildings('elec_kWh', period) / buildings('total_energy_kwh', period) * 100
+        return np.round(elec_percent, 2)
+
+
+class perc_gas_kwh(Variable):
+    value_type = float
+    entity = Building
+    definition_period = ETERNITY
+    label = "The percentage of a building's energy consumption that is derived from its electricity consumption, expressed as a percentage of total energy use to 2 decimal places"
+
+    def formula(buildings, period, parameters):
+        gas_percent = buildings('gas_kWh', period) / buildings('total_energy_kwh', period) * 100
+        return np.round(gas_percent, 2)
+
+
+class perc_diesel_kwh(Variable):
+    value_type = float
+    entity = Building
+    definition_period = ETERNITY
+    label = "The percentage of a building's energy consumption that is derived from its diesel consumption, expressed as a percentage of total energy use to 2 decimal places"
+
+    def formula(buildings, period, parameters):
+        diesel_percent = buildings('diesel_kWh', period) / buildings('total_energy_kwh', period) * 100
+        return np.round(diesel_percent, 2)
+
+
+class perc_coal_kwh(Variable):
+    value_type = float
+    entity = Building
+    definition_period = ETERNITY
+    label = "The percentage of a building's energy consumption that is derived from its oil consumption, expressed as a percentage of total energy use to 2 decimal places"
+
+    def formula(buildings, period, parameters):
+        coal_percent = buildings('coal_kWh', period) / buildings('total_energy_kwh', period) * 100
+        return np.round(coal_percent, 2)
+
+
+class offices_benchmark_star_rating(Variable):
+    value_type = float
+    entity = Building
+    definition_period = ETERNITY
+    label = 'The benchmark star rating used for calculating Benchmark' \
+            ' Electricity and Benchmark Gas consumption. Note that as the' \
+            ' Offices Reverse Calculator can only calculate to the nearest ' \
+            ' half star, this is rounded down to the nearest half star, as' \
+            ' prescribed in Step 3 of Calculation Method 4.'
+
+    def formula(buildings, period, parameters):
+        input_star_rating = buildings('benchmark_star_rating', period)
+        rounded_to_whole_star = np.floor(input_star_rating)
+        condition_half_star = (input_star_rating - rounded_to_whole_star) <= 0.5
+        return where(condition_half_star, rounded_to_whole_star, (rounded_to_whole_star + 0.5))
 
 
 class nabers_adjusted_hours(Variable):
@@ -382,3 +536,171 @@ class coefficient_B(Variable):
 			 -0.0166656, -0.0359809, -0.0310451, -0.0151039, -0.0341818, -0.0270617,
 			 -0.0148, -0.0444444, -0.0223, -0.02381, -0.05, -0.04545]
             )
+
+
+class NGEmax (Variable):
+    value_type = float
+    entity = Building
+    definition_period = ETERNITY
+    label = 'Normalised greenhouse emissions for relevant NABERS whole' \
+            'building - only applicable under 5 stars'
+
+    def formula(buildings, period, parameters):
+        A = buildings('coefficient_A', period)
+        B = buildings('coefficient_B', period)
+        return ((buildings('benchmark_star_rating', period) - 0.499999 - A) / B)
+
+
+class base_building_GEmax (Variable):
+    value_type = float
+    entity = Building
+    definition_period = ETERNITY
+    label = "GEmax value for base building ratings."
+
+    def formula(buildings, period, parameters):
+        NGEmax = buildings('NGEmax', period)
+        f_base_building = buildings('f_base_building', period)
+        GEclimcorr = buildings('GEclimcorr', period)
+        return (NGEmax / f_base_building - GEclimcorr)
+
+
+class tenancy_GEmax (Variable):
+    value_type = float
+    entity = Building
+    definition_period = ETERNITY
+    label = "GEmax value for tenancy ratings."
+
+    def formula(buildings, period, parameters):
+        NGEmax = buildings('NGEmax', period)
+        f_tenancy = buildings('f_tenancy', period)
+        GEClimcorr_tenancy = buildings('GEClimcorr_tenancy', period)
+        return (NGEmax / f_tenancy - GEClimcorr_tenancy)
+
+
+class GEwholemax (Variable):
+    value_type = float
+    entity = Building
+    definition_period = ETERNITY
+    label = "Greenhouse emissions for relevant NABERS whole building - applicable above 5 stars"
+
+    def formula(buildings, period, parameters):
+        condition_GEwholemax_star_rating = buildings('benchmark_star_rating', period) > 5
+        return where(condition_GEwholemax_star_rating, 0
+        , (buildings('NGEmax', period) - buildings('f_base_building', period) * buildings('GEclimcorr', period) - buildings('f_tenancy', period) * buildings('GEClimcorr_tenancy', period)) * 2 / (buildings('f_base_building', period) + buildings('f_tenancy', period)))
+
+
+class NGE_5star_original_rating (Variable):
+    value_type = float
+    entity = Building
+    definition_period = ETERNITY
+    label = "original normalised greenhouse emissions at 5 star for use for > 5 star ratings"
+
+    def formula(buildings, period, parameters):
+        A = buildings('coefficient_A', period)
+        B = buildings('coefficient_B', period)
+        return ((5 - A - 0.499999) / B)
+
+
+class GE_5star_original_rating (Variable):
+    value_type = float
+    entity = Building
+    definition_period = ETERNITY
+    label = "original required maximum greenhouse emissions at 5 star rating"
+
+    def formula(buildings, period, parameters):
+        state = buildings('building_state_location', period)
+        NGE_rating = buildings('NGE_5star_original_rating', period)
+        f_bb = buildings('f_base_building', period)
+        GEclimcorr = buildings('GEclimcorr', period)
+        f_ten = buildings('f_tenancy', period)
+        GEClimcorr_ten = buildings('GEClimcorr_tenancy', period)
+        rating_type = buildings('rating_type', period)
+        return select(
+            [rating_type == "base_building"
+            , rating_type == "tenancy"
+            , rating_type == "whole_building"],
+            [NGE_rating / f_bb - GEclimcorr
+            , NGE_rating / f_ten - GEClimcorr_ten
+            , (NGE_rating - f_bb * GEclimcorr - f_ten * GEClimcorr_ten)
+             * 2 / (f_bb+f_ten)]
+            )
+
+
+class GE_25_percent_reduction(Variable):
+    value_type = float
+    entity = Building
+    definition_period = ETERNITY
+    label = "25% reduction of required maximum greenhouse emissions for use in 5.5 star rating following original rating system"
+
+    def formula(buildings, period, parameters):
+        GE_5_stars = buildings('GE_5star_original_rating', period)
+        return GE_5_stars * 0.75
+
+
+class GE_50_percent_reduction(Variable):
+    value_type = float
+    entity = Building
+    definition_period = ETERNITY
+    label = "50% reduction of required maximum greenhouse emissions for use in 6 star rating following original rating system"
+
+    def formula(buildings, period, parameters):
+        GE_5_stars = buildings('GE_5star_original_rating', period)
+        return GE_5_stars * 0.5
+
+
+class office_maximum_electricity_consumption(Variable):
+    value_type = float
+    entity = Building
+    definition_period = ETERNITY
+    label = "output of the NABERS whole building reverse calculator - the maximum electricity consumption allowable for the relevant NABERS rated building"  # need to figure out how to round this to an integer
+
+    def formula(buildings, period, parameters):
+        nla = buildings('net_lettable_area', period)
+        benchmark = buildings('benchmark_star_rating', period)
+        SGEelec = buildings('SGEelec', period)
+        SGEgas = buildings('SGEgas', period)
+        SGEcoal = buildings('SGEcoal', period)
+        SGEoil = buildings('SGEoil', period)
+        bb_GEmax = buildings('base_building_GEmax', period)
+        ten_GEmax = buildings('tenancy_GEmax', period)
+        GEwholemax = buildings('GEwholemax', period)
+        GE_25_perc = buildings('GE_25_percent_reduction', period)
+        GE_50_perc = buildings('GE_50_percent_reduction', period)
+        perc_gas = buildings('perc_gas_kwh', period)
+        perc_elec = buildings('perc_elec_kwh', period)
+        perc_coal = buildings('perc_coal_kwh', period)
+        perc_diesel = buildings('perc_diesel_kwh', period)
+        rating_type = buildings('rating_type', period)
+
+        def consumption(numerator):
+            return np.round((numerator * nla) / (SGEelec + perc_gas / perc_elec * SGEgas + perc_coal / perc_elec * SGEcoal + perc_diesel / perc_elec * SGEoil), 9)
+
+        return select(
+            [benchmark <= 5 and rating_type == "base_building"
+            , benchmark <= 5 and rating_type == "whole_building"
+            , benchmark <= 5 and rating_type == "tenancy"
+            , benchmark == 5.5 and rating_type == "whole_building"
+            , benchmark == 5.5 and rating_type == "base_building"
+            , benchmark == 5.5 and rating_type == "tenancy"
+            , benchmark == 6 and rating_type == "whole_building"
+            , benchmark == 6 and rating_type == "base_building"
+            , benchmark == 6 and rating_type == "tenancy"],
+             [consumption(bb_GEmax)
+             , consumption(GEwholemax)
+             , consumption(ten_GEmax)
+             , consumption(GE_25_perc)
+             , consumption(GE_25_perc)
+             , consumption(GE_25_perc)
+             , consumption(GE_50_perc)
+             , consumption(GE_50_perc)
+             , consumption(GE_50_perc)])
+
+
+class office_maximum_gas_consumption(Variable):
+    value_type = float
+    entity = Building
+    definition_period = ETERNITY
+    label = "output of the NABERS whole building reverse calculator - the maximum electricity consumption allowable for the relevant NABERS rated building"
+
+    def formula(buildings, period, parameters):
+        return (buildings('perc_gas_kwh', period) / buildings('perc_elec_kwh', period) * buildings('office_maximum_electricity_consumption', period) * 3.6)
