+++ openfisca_nsw_ess_nabers/variables/energy_saving_scheme/general_ESS/clause_6.5_ESC_calculation.py
@@ -3,6 +3,7 @@ from openfisca_core.model_api import *
 # Import the Entities specifically defined for this tax and benefit system
 from openfisca_nsw_base.entities import *
 import math
+import numpy as np
 
 
 class number_of_certificates(Variable):
@@ -18,4 +19,4 @@ class number_of_certificates(Variable):
         elec_cert_conversion_factor = parameters(period).energy_saving_scheme.electricity_certificate_conversion_factor
         gas_savings = buildings('gas_savings', period)
         gas_cert_conversion_factor = parameters(period).energy_saving_scheme.gas_certificate_conversion_factor
-        return math.floor((elec_savings * elec_cert_conversion_factor) + (gas_savings * gas_cert_conversion_factor))
+        return math.floor(np.round((elec_savings * elec_cert_conversion_factor) + (gas_savings * gas_cert_conversion_factor), 1))
